Conflicts:
	README.md
	src/game/BattleGround/BattleGroundHandler.cpp
	src/game/SpellEffects.cpp
	src/game/World.cpp

+++ src/shared/Threading.h
@@ -19,6 +19,7 @@
 #ifndef THREADING_H
 #define THREADING_H
 
+#include "Common.h"
 #include <ace/Thread.h>
 #include <ace/TSS_T.h>
 #include "ace/Atomic_Op.h"
@@ -39,7 +40,7 @@ namespace ACE_Based
                     delete this;
             }
         private:
-            ACE_Atomic_Op<ACE_Thread_Mutex, long> m_refs;
+            ACE_Atomic_Op<ACE_Thread_Mutex, uint32> m_refs;
     };
 
     enum Priority
